requirements.txt
# store
sqlalchemy~=2.0.31
PyMySQL~=1.1.1

# utils
markdown~=3.6
Pygments~=2.18.0
pytz>=2024.1
jinja2~=3.1.4
psutil~=6.0.0
unicategories~=0.1.2

# logic
pyzmq>=26.0.3
pyopenssl>=24.1.0

# api
sanic~=23.12
sanic-ext~=23.12
httpx[http2]~=0.27.0
pyjwt~=2.8.0
cryptography>=42.0.8
websockets>=11.0 # websockets.connections renamed to websockets.protocol since then

# admin
flask~=3.0.3
flask-sqlalchemy~=3.1.1
flask-admin~=1.6.1
Flask-Migrate~=4.0.4
wtforms~=3.1.2
markupsafe~=2.1.1
gevent>=21.12.0

# mail
base58~=2.1.1
pyCryptodome~=3.18.0
sanic_mail==0.0.3
